move Game and TitleScreen into a screens subfolder in components
